requirements.txt
streamlit>=1.30.0
streamlit-folium>=0.18.0
folium>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.43.0
plotly>=5.18.0
statsmodels>=0.14.0
requests>=2.31.0
scipy>=1.11.0
matplotlib>=3.7.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
daftlistings>=2.0.4
google-generativeai>=0.7.0
